refactor delete

+++ src/projects/deleteProject.ts
@@ -1,12 +1,21 @@
 import { Response, Request } from "express";
 import { StatusCodes } from "http-status-codes";
+
 import { projectModel } from "../../models/project.model";
+import userModel from "../../models/user.model";
 
 const deleteProject = async (req: Request, res: Response) => {
-  // TODO only owner should has access to delete project
-  const project = await projectModel.findByIdAndDelete(req.params.id);
+  const project = await projectModel.findById(req.params.id);
   if (!project) return res.status(StatusCodes.NOT_FOUND).send("Project not found");
 
+  const { _id: id } = req.userInfo;
+  const { userId, mentorId } = project;
+  const userOK = userId?.toString() === id.toString();
+  const mentorOK = mentorId?.toString() === id.toString();
+  if (!userOK && !mentorOK)
+    return res.status(StatusCodes.FORBIDDEN).send(`You are not authorized to perform this operation`);
+
+  await projectModel.findByIdAndDelete(req.params.id);
   return res.status(StatusCodes.OK).send(`Project with id: ${req.params.id} has been deleted`);
 };
